requirements.txt
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0
opencv-python==4.8.0
pillow==10.0.0